IBX-538: Renamed wrong namespace in RepositoryInstaller tests
