+++ tests/bundle/RepositoryInstaller/DependencyInjection/Compiler/InstallerTagPassTest.php
@@ -6,7 +6,7 @@
  */
 declare(strict_types=1);
 
-namespace Ibexa\Tests\Bundle\Installer\DependencyInjection\Compiler;
+namespace Ibexa\Tests\Bundle\RepositoryInstaller\DependencyInjection\Compiler;
 
 use Ibexa\Bundle\RepositoryInstaller\Command\InstallPlatformCommand;
 use Ibexa\Bundle\RepositoryInstaller\DependencyInjection\Compiler\InstallerTagPass;
